pyproject.toml
[tool.poetry]
name = "brigid"
version = "0.12.4"
description = "Static site generator."
readme = "README.md"
repository = "https://github.com/Tiendil/brigid"
authors = ["Aliaksei Yaletski (Tiendil) <a.eletsky@gmail.com>"]
license = "BSD-3-Clause"
keywords = ["blog", "blog-engine", "site", "markdown", "cms", "content management system", "personal site"]
classifiers = [
        "Environment :: Web Environment",
        "Framework :: AsyncIO",
        "Framework :: FastAPI",
        "Intended Audience :: End Users/Desktop",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: BSD License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: Implementation :: CPython",
        "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
        "Topic :: Office/Business :: News/Diary",
        "Topic :: Text Processing :: Markup",
        "Topic :: Text Processing :: Markup :: HTML",
        "Topic :: Text Processing :: Markup :: Markdown",
        "Topic :: Internet :: WWW/HTTP",
        "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
        "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
        "Topic :: Internet :: WWW/HTTP :: WSGI :: Server",
        "Topic :: Internet :: WWW/HTTP :: HTTP Servers"]

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "0.115.*"
structlog = "24.1.*"

toml = "0.10.*"

pydantic = "2.10.*"
pydantic-settings = "2.6.*"
pillow = "10.2.*"

jinja2 = "3.1.*"
markdown = "3.5.*"
pymdown-extensions = "10.7.*"
python-frontmatter = "1.1.*"
feedgenerator = "2.1.*"
beautifulsoup4 = "4.12.*"
html5lib = "1.1.*"

sentry-sdk = "1.39.*"
typer = "0.9.*"

[tool.poetry.group.dev.dependencies]
uvicorn = {version = "0.27.*", extras = ["standard"]}

changy = "0.*"

pytest = "7.4.*"
pytest-asyncio = "0.21.*"
pytest-mock = "3.11.*"

httpx = "0.27.*"

isort = "5.13.*"
black = "24.2.*"

flake8 = "7.0.*"
flake8-docstrings = "1.7.*"
flake8-functions = "0.0.*"
flake8-annotations-complexity = "0.0.*"
flake8-cognitive-complexity = "0.1.*"
flake8-pytest = "1.4.*"
flake8-bandit = "4.1.*"
flake8-absolute-import = "1.0.*"
flake8-print = "5.0.*"
flake8-pyproject = "1.2.*"
flake8-eradicate = "1.5.*"
autoflake = "2.3.*"

mypy = "1.8.*"

types-toml = "0.10.*"
types-Markdown = "3.5.*"
types-Pillow = "10.2.*"
types-beautifulsoup4 = "4.12.*"

[tool.poetry.scripts]
brigid = "brigid.cli.__main__:app"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119

[tool.autoflake]
in_place = true
remove_all_unused_imports = true
remove_unused_variables = true
recursive = true

[tool.flake8]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "D200",
    "D400",
    "S101",
    "S311",
    "W503",
    "CFQ004",
    "E203",  # blake formats [x:y] with spaves
]
max-line-length = 119

[[tool.mypy.overrides]]
module = "feedgenerator.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "frontmatter.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pymdownx.*"
ignore_missing_imports = true


[tool.setuptools]
packages = ["brigid"]

[project]
name = "brigid"
version = "0.12.4"
